Re-run the JS if the Theme changes
